refactor(dto): 중복된 dto 디렉토리 수정
