--- src/main/java/wooteco/subway/dto/StationRequest.java
@@ -1,22 +1,0 @@
-package wooteco.subway.dto;
-
-import wooteco.subway.domain.Station;
-
-public class StationRequest {
-    private String name;
-
-    public StationRequest() {
-    }
-
-    public StationRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Station toStation() {
-        return new Station(name);
-    }
-}
