+++ src/main/java/wooteco/subway/service/PathService.java
@@ -7,10 +7,10 @@ import wooteco.subway.domain.Path;
 import wooteco.subway.domain.PathGenerator;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.PathResponse;
 import wooteco.subway.jgraph.JGraphPathGenerator;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
+import wooteco.subway.service.dto.PathResponse;
 
 @Service
 @Transactional
